Create separate component for FancySlider
