+++ web/frontend/page_mitigation_scenarios/MitigationCalculator.tsx
@@ -1,7 +1,7 @@
-import * as chroma from "chroma-js";
-import * as d3 from "d3";
-import * as React from "react";
+import * as d3 from 'd3';
+import * as React from 'react';
 
+import FancySlider from '../components/FancySlider';
 import {
   calculateHighCompliance,
   calculateLowCompliance,
@@ -9,246 +9,12 @@ import {
   Measure,
   MeasureGroup,
   range,
-} from "./measures";
+} from './measures';
 
 //let scale = chroma.scale("PuBu");
 // let scale = chroma
 //   .scale(["black", "red", "yellow", "white"])
 //   .correctLightness();
-
-function p(v: number): string {
-  return `${(100 * v).toFixed(5)}%`;
-}
-
-function calculateBackground(
-  mean: number,
-  sd: number,
-  min: number,
-  max: number,
-  thumbWidth: string,
-  scale: chroma.Scale
-): string {
-  function getColor(z: number) {
-    return scale(Math.exp((z * z) / -2)).css();
-  }
-
-  let backgrounds = [];
-  const f = (loc: number) => getColor((loc - mean) / sd);
-  backgrounds.push(`linear-gradient(
-    to right, ${f(min)} 49%, ${f(max)} 51%)`);
-
-  let stops: Array<string> = [];
-  let addStop = (loc: number, z: number) => {
-    let offset = (loc - min) / (max - min);
-    stops.push(`${getColor(z)} ${(offset * 100).toFixed(2)}%`);
-  };
-
-  for (let z = -4; z < 4; z += 0.25) {
-    addStop(mean + z * sd, z);
-  }
-
-  let rulerGradient = `linear-gradient(to right, ${stops.join(",")})`;
-  backgrounds.push(
-    `no-repeat ${rulerGradient} 
-      calc(${thumbWidth}/2) 0 
-    / calc(100% - ${thumbWidth}) auto`
-  );
-
-  function addTicks(
-    interval: number,
-    height: number,
-    color: string,
-    color2: string,
-    pos?: number
-  ) {
-    let w = 1;
-    let firstTick = pos ?? Math.ceil(min / interval) * interval;
-    let offset = p((firstTick - min) / (max - min));
-
-    let tickSpace = p(interval / (max - min));
-    // let tickGradient = `repeating-linear-gradient(to right,
-    //   ${color}, ${color} 1px,
-    //   transparent 1px, transparent ${tickSpace}
-    // )`;
-    let tickGradient = `${
-      pos === undefined ? "repeating-linear-gradient" : "linear-gradient"
-    }(to right, 
-      transparent calc(${offset}), ${color} calc(${offset}),
-      ${color} calc(${offset} + ${w}px), ${color2} calc(${offset} + ${w}px), 
-      ${color2} calc(${offset} + ${2 * w}px), transparent calc(${offset} + ${
-      2 * w
-    }px)
-      ${
-        pos !== undefined ? "" : `, transparent calc(${offset} + ${tickSpace})`
-      } 
-    )`;
-    backgrounds.push(
-      `no-repeat ${tickGradient} 
-        calc(${thumbWidth} / 2) bottom 
-      / calc(100% - ${thumbWidth}) ${p(height)}`
-    );
-  }
-  let dark = scale(1).desaturate().css();
-  let light = scale(0.5).desaturate().css();
-
-  if (max < 2) {
-    addTicks(0.05, 0.2, dark, light);
-    addTicks(0.1, 0.5, dark, light);
-  } else if (max - min < 4) {
-    addTicks(0.1, 0.2, dark, light);
-    addTicks(1, 0.5, dark, light);
-  } else {
-    addTicks(0.5, 0.2, dark, light);
-    addTicks(1, 0.5, dark, light);
-  }
-  addTicks(1, 1, dark, scale(0.4).desaturate().css(), 1);
-
-  return backgrounds.reverse().join(", ");
-}
-
-namespace FancySlider {
-  export interface Props {
-    mean: number;
-    sd: number;
-    disabled?: boolean;
-    initial?: number;
-    row: number;
-    step: number;
-    min: number;
-    max: number;
-    value: number;
-    scale?: chroma.Scale;
-    onChange?: (value: number) => void;
-    format?: "percentage" | "absolute";
-  }
-}
-
-function FancySlider({
-  onChange,
-  mean,
-  sd,
-  initial: propInitial,
-  min: propMin,
-  max: propMax,
-  step,
-  row,
-  disabled: propDisabled,
-  value: propValue,
-  format: propFormat,
-  scale,
-}: FancySlider.Props) {
-  let format =
-    propFormat == "percentage"
-      ? (x: number) => d3.format("+.0%")(x - 1)
-      : (x: number) => x.toFixed(1);
-  const [initial] = React.useState(propInitial ?? mean);
-  let disabled = propDisabled ?? false;
-  let min = Math.floor(propMin / step) * step;
-  let max = Math.ceil(propMax / step) * step;
-  let value = Math.round(propValue / step) * step;
-
-  const background = React.useMemo(() => {
-    return calculateBackground(
-      mean,
-      Math.abs(sd),
-      propMin,
-      propMax,
-      onChange !== undefined ? "var(--thumb-width)" : "3px",
-      scale ?? chroma.scale(["rgb(212, 212, 157)", "rgb(102, 102, 66)"])
-    );
-  }, [onChange, mean, sd, min, max]);
-
-  let input: JSX.Element;
-  if (onChange) {
-    let lin: (x: number) => number = (x) => x;
-    let inv: (x: number) => number = (x) => x;
-    let aff: (x: number) => number = (x) => x;
-    if (propFormat == "percentage") {
-      lin = (x) => 100 * (x - 1);
-      inv = (x) => 1 + x / 100;
-      aff = (x) => 100 * x;
-    }
-
-    input = (
-      <div className="input-group">
-        <div className="input-group-prepend">
-          <button
-            type="button"
-            className="btn btn-secondary"
-            onClick={() => onChange(initial)}
-          >
-            ↻
-          </button>
-        </div>
-        <input
-          className="form-control"
-          min={lin(min)}
-          max={lin(max)}
-          step={aff(step)}
-          type="number"
-          onChange={(evt) => onChange(inv(+evt.target.value))}
-          value={
-            propFormat == "percentage"
-              ? lin(value).toFixed(0)
-              : value.toFixed(-Math.floor(Math.log10(step)))
-          }
-        ></input>
-        {propFormat == "percentage" ? (
-          <div className="input-group-append">
-            <span className="input-group-text">%</span>
-          </div>
-        ) : (
-          ""
-        )}
-      </div>
-    );
-  } else {
-    input = <b>{value.toFixed(-Math.floor(Math.log10(step)))}</b>;
-  }
-
-  return (
-    <>
-      <div
-        className="slider"
-        style={{
-          gridColumn: 3,
-          gridRow: row,
-          filter: disabled ? "brightness(50%)" : "none",
-        }}
-      >
-        <span className="ruler-label">{format(min)}</span>
-        <input
-          className="ruler measure-slider"
-          type="range"
-          disabled={onChange === undefined}
-          value={propValue ?? undefined}
-          min={propMin}
-          max={propMax}
-          step="any"
-          onChange={onChange ? (evt) => onChange(+evt.target.value) : undefined}
-          style={{
-            // @ts-ignore
-            "--ruler-background": background,
-          }}
-        ></input>
-        <span className="ruler-label">{format(max)}</span>
-      </div>
-      <div
-        key={`value-${row}`}
-        className="value"
-        style={{
-          // @ts-ignore
-          gridColumn: 4,
-          gridRow: row,
-          filter: disabled ? "brightness(50%)" : "none",
-          justifySelf: "end",
-        }}
-      >
-        {input}
-      </div>
-    </>
-  );
-}
 
 type LabeledCheckboxProps = {
   id: string;
